refactor: Created dedicated frontend package, with its own sub package for screens
